Fix evaltool build errors 270rc (#1056)
* Fix Windows build errors caused by refactor

* Remove ../../SDK/src from #includes

+++ src/TcpServerPortFactory.h
@@ -133,7 +133,7 @@ protected:
     void stopListening();
 
     int getClientConnectionCount() {
-        return knownSockets.size();
+        return (int)knownSockets.size();
     }
 
     std::vector<socket_entry_t> getClientSockets() {
